+++ graphql/types/User.js
@@ -19,4 +19,32 @@ module.exports=
 
 type Query{
     users: [User]
-}`;
+}
+type Mutation {
+    addUser(input: UserInput): User
+}
+
+input UserInput {
+    username: String,
+    password: String,
+    first_name: String,
+    last_name: String,
+    genre: String,
+    birth_date: Date,
+    phone: String,
+    address: AddressInput, 
+    identification: IdentificationInput,
+    email: String
+}
+input AddressInput {
+    country: String,
+    city: String,
+    street: String,
+    state: String,
+    aditional_info: String
+}
+input IdentificationInput{
+    identification_type: String,
+    identification_num: String
+}
+`;
